Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,7 +2,7 @@
 #include<stdio.h>
 int main()
 {
-    float r,area;
+    double r,area;
     scanf("%.2f",&r);
     area = 3.14*r*r;
     printf("Area of circle %.2f",area);
